requirements.txt
aerich==0.4.3; python_version > '3.6'
aPRAW==0.6.8a0
discord.py==1.6.0
fastapi==0.63.0
jishaku==1.20.0.220
python-dotenv==0.15.0
PyYAML==5.3.1
tortoise-orm[asyncpg]==0.16.19; python_version > '3.6'
async-exit-stack; python_version < '3.7'
async-generator; python_version < '3.7'
